+++ OmniProxy.cpp
@@ -21,6 +21,7 @@
 #include "defines.hpp"
 
 #include <cstdarg>
+#include <cstdint>
 
 #include <unistd.h>
 #include <sys/mman.h>
@@ -1233,7 +1234,7 @@ std::string OmniProxy::
 toString (std::ostringstream& oss)
 {
    oss << "{OmniProxy: pDevice_d = " << *pDevice_d
-       << ", pHeadersRoot_d = 0x" << std::hex << (int)pHeadersRoot_d << std::dec
+               << ", pHeadersRoot_d = 0x" << std::hex << (uintptr_t)pHeadersRoot_d << std::dec
        << ", pHeadersCurrent_d = 0x" << pHeadersCurrent_d
        << ", pHeaderCurrent_d = 0x" << pHeaderCurrent_d
        << ", fp_d = 0x" << fp_d
